QCMPLUS-31 update User models and UserService

+++ src/main/java/com/qcmplus/qcmplus/model/Answers.java
@@ -1,6 +1,13 @@
 package com.qcmplus.qcmplus.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,7 +19,7 @@ public class Answers {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "answer_text")
     private String answerText;
