+++ src/main/java/com/qcmplus/qcmplus/model/Quizzes.java
@@ -1,6 +1,12 @@
 package com.qcmplus.qcmplus.model;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -14,7 +20,7 @@ public class Quizzes {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long quiz_id;
+    private Integer quiz_id;
 
     @Column(name = "title")
     private String title;
